requirements.txt
streamlit
aiohttp
openai
yfinance
requests
python-dotenv 